Renamed Icon files, added info icon and subheader url config
