+++ apps/verification-portal/src/lib/components/icons/HexagonIcon.svelte
