+++ apps/verification-portal/src/lib/components/icons/InfoIcon.svelte
@@ -1,0 +1,19 @@
+<script lang="ts">
+  export let className: string = '';
+</script>
+
+<svg
+  class={className}
+  aria-hidden="true"
+  xmlns="http://www.w3.org/2000/svg"
+  fill="none"
+  viewBox="0 0 20 20"
+>
+  <path
+    stroke="currentColor"
+    stroke-linecap="round"
+    stroke-linejoin="round"
+    stroke-width="2"
+    d="M8 9h2v5m-2 0h4M9.408 5.5h.01M19 10a9 9 0 1 1-18 0 9 9 0 0 1 18 0Z"
+  />
+</svg>
